Silence compiler warnings
test/pi_table.cpp:60:16: error:
      comparison of integers of different signs: 'uint64_t'
      (aka 'unsigned long') and 'int64_t' (aka 'long')
[-Werror,-Wsign-compare]
  while (prime < pi.size())
         ~~~~~ ^ ~~~~~~~~~
/home/travis/build/kimwalisch/primecount/test/pi_table.cpp:63:21:
error:
      comparison of integers of different signs: 'int64_t' (aka 'long')
and
      'uint64_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
    check(pi[prime] == count);
          ~~~~~~~~~ ^  ~~~~~
/home/travis/build/kimwalisch/primecount/test/pi_table.cpp:72:17:
error:
      comparison of integers of different signs: 'int64_t' (aka 'long')
and
      'uint64_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
    check(pi[n] == primesieve::count_primes(0, n));

+++ test/pi_table.cpp
@@ -54,8 +54,8 @@ int main()
   }
 
   primesieve::iterator it;
-  uint64_t prime = it.next_prime();
-  uint64_t count = 1;
+  int64_t prime = it.next_prime();
+  int64_t count = 1;
 
   while (prime < pi.size())
   {
@@ -69,7 +69,7 @@ int main()
   {
     int n = dist(gen) % pi.size();
     cout << "pi(" << n << ") = " << pi[n];
-    check(pi[n] == primesieve::count_primes(0, n));
+    check(pi[n] == (int64_t) primesieve::count_primes(0, n));
   }
 
   cout << endl;
